extract Java5 plugin into VMPluginBase class
