+++ src/main/org/codehaus/groovy/ast/decompiled/FormalParameterParser.java
@@ -20,7 +20,7 @@ package org.codehaus.groovy.ast.decompiled;
 
 import org.codehaus.groovy.ast.ClassNode;
 import org.codehaus.groovy.ast.GenericsType;
-import org.codehaus.groovy.vmplugin.v5.Java5;
+import org.codehaus.groovy.ast.tools.GenericsUtils;
 import org.objectweb.asm.Opcodes;
 import org.objectweb.asm.signature.SignatureVisitor;
 
@@ -49,9 +49,9 @@ abstract class FormalParameterParser extends SignatureVisitor {
 
     protected void flushTypeParameter() {
         if (currentTypeParameter != null) {
-            ClassNode ref = Java5.configureTypeVariableReference(currentTypeParameter);
+            ClassNode ref = GenericsUtils.configureTypeVariableReference(currentTypeParameter);
             ClassNode[] boundNodes = parameterBounds.toArray(new ClassNode[parameterBounds.size()]);
-            typeParameters.add(Java5.configureTypeVariableDefinition(ref, boundNodes));
+            typeParameters.add(GenericsUtils.configureTypeVariableDefinition(ref, boundNodes));
 
             currentTypeParameter = null;
             parameterBounds.clear();
