+++ src/main/org/codehaus/groovy/vmplugin/v5/package.html
@@ -23,6 +23,9 @@
         <p>
             Java 5 specific classes.
         </p>
+        <p>
+            This package deprecated and is being kept for compatibility.
+        </p>
     </body>
 </html>
